update MACFin user after every successful password rotation
This ensures the automated sheet and the MACFin sheet have the
same password for each user before uploading the file to S3.
If the upload fails, the file stored in S3 has, at most, 1 user
with an invalid password, and both sheets are consistent.

The MACFin team can manually update this password in their sheet in the S3 file.
The MAC-FC team can manually remove this user from the automated sheet.
The next time the tool runs, the automated sheet picks up the user with
the new password, during synchronization.